+++ lib/widgets/journal/entry_details/workout_summary.dart
@@ -30,8 +30,6 @@ class WorkoutSummary extends StatelessWidget {
       }
     });
 
-    debugPrint(items.toString());
-
     return Padding(
       padding: const EdgeInsets.all(8.0),
       child: Column(